refactor packing & common content on VarCons & i3s Prod
