+++ .idea/encodings.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/binscale-common/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/binscale-common/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/multipleConsumers/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/multipleConsumers/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/restructuredbinpack/src/main/java" charset="UTF-8" />
