+++ IntegrationController/src/main/java/fr/unice/scale/latencyaware/controller/bin_pack/BinPackLag.java
@@ -1,5 +1,12 @@
-package fr.unice.scale.latencyaware.controller;
+package fr.unice.scale.latencyaware.controller.bin_pack;
 
+import fr.unice.scale.latencyaware.common.config.KafkaConsumerConfig;
+import fr.unice.scale.latencyaware.controller.ArrivalProducer;
+import fr.unice.scale.latencyaware.controller.entity.Consumer;
+import fr.unice.scale.latencyaware.controller.entity.Partition;
+import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.kubernetes.client.KubernetesClientBuilder;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -8,57 +15,92 @@ import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Properties;
 
-
-public class BinPackState {
+public class BinPackLag {
     //TODO give fup and fdown as paramters to the functions.
-    private static final Logger log = LogManager.getLogger(BinPackState.class);
-    public static int size = Integer.valueOf(System.getenv("INIT_SIZE"));
-    public Instant LastUpScaleDecision = Instant.now();
-    
-    static double wsla = Double.valueOf(System.getenv("WSLA"));
 
-    static double mu = Double.valueOf(System.getenv("MU"));
+    private static final Logger log = LogManager.getLogger(BinPackLag.class);
 
-    static String action = "none";
+    public static Instant LastUpScaleDecision = Instant.now();
 
-    static List<Consumer> assignment = new ArrayList<Consumer>();
-    static List<Consumer> currentAssignment = assignment;
-    static List<Consumer> tempAssignment = assignment;
+    public static double wsla = Double.valueOf(System.getenv("WSLA"));
 
-    public  static void scaleAsPerBinPack() {
-        action = "none";
-        log.info("Currently we have this number of consumers group {} {}","testgroup1", size );
-        int neededsize = binPackAndScale();
-        log.info("We currently need the following consumers for group1 (as per the bin pack) {}", neededsize);
-        int replicasForscale = neededsize - size;
-        if (replicasForscale > 0) {
-            action = "up";
-            //TODO IF and Else IF can be in the same logic
-            log.info("We have to upscale  group1 by {}", replicasForscale);
-            //currentAssignment = assignment;
-            return;
+    public static double rebTime= Double.valueOf(System.getenv("REB_TIME"));
 
-        } else {
-            int neededsized = binPackAndScaled();
-            int replicasForscaled = size - neededsized;
-            if (replicasForscaled > 0) {
-               action = "down";
-                log.info("We have to downscale  group by {} {}", "testgroup1", replicasForscaled);
-                //currentAssignment = assignment;
-                return;
-            }
+    public static List<Consumer> assignment = new ArrayList<Consumer>();
+    public static List<Consumer> currentAssignment = assignment;
+
+    private static KafkaConsumer<byte[], byte[]> metadataConsumer;
+
+    public static double mu = Double.valueOf(System.getenv("MU"));
+
+    static {
+        float fup = Float.valueOf(System.getenv("FUP"));
+        currentAssignment.add(new Consumer("0", (long) (mu * wsla * fup),
+                mu * fup));
+        for (Partition p : ArrivalProducer.topicpartitions) {
+            currentAssignment.get(0).assignPartition(p);
+        }
+    }
+
+
+    public static  void scaleAsPerBinPack() {
+
+        log.info("Currently we have this number of consumers group {} {}","testgroup1", BinPackState.size );
+
+        for (int i = 0; i < ArrivalProducer.numberpartitions; i++) {
+            ArrivalProducer.topicpartitions.get(i).setLag(ArrivalProducer.topicpartitions.get(i).getLag()
+                    + (long) ((ArrivalProducer.totalArrivalrate * rebTime)/(ArrivalProducer.numberpartitions * 1.0)));
         }
 
-
-        if (assignmentViolatesTheSLA2()) {
-            action = "REASS";
+        if (BinPackState.action.equals("up") || BinPackState.action.equals("REASS")) {
+            int neededsize = binPackAndScale();
+            log.info("We currently need the following consumers for group1 (as per the bin pack) {}", neededsize);
+            int replicasForscale = neededsize - BinPackState.size;
+            if (replicasForscale > 0) {
+                //TODO IF and Else IF can be in the same logic
+                log.info("We have to upscale  group1 by {}", replicasForscale);
+                BinPackState.size = neededsize;
+                LastUpScaleDecision = Instant.now();
+                currentAssignment = assignment;
+                try (final KubernetesClient k8s = new KubernetesClientBuilder().build() ) {
+                k8s.apps().deployments().inNamespace("default").withName("latency").scale(neededsize);
+                log.info("I have Upscaled group {} you should have {}", "testgroup1", neededsize);
+            }
+            } else if (replicasForscale == 0) {
+                if (metadataConsumer == null) {
+                    KafkaConsumerConfig config = KafkaConsumerConfig.fromEnv();
+                    Properties props = KafkaConsumerConfig.createProperties(config);
+                    props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG,
+                            "org.apache.kafka.common.serialization.StringDeserializer");
+                    props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG,
+                            "org.apache.kafka.common.serialization.StringDeserializer");
+                    metadataConsumer = new KafkaConsumer<>(props);
+                    metadataConsumer.enforceRebalance();
+                }
+                currentAssignment = assignment;
+                metadataConsumer.enforceRebalance();
+            }
+        } else if (BinPackState.action.equals("down")) {
+            int neededsized = binPackAndScaled();
+            int replicasForscaled =  BinPackState.size - neededsized;
+            if (replicasForscaled <  BinPackState.size ) {
+                log.info("We have to downscale  group by {} {}", "testgroup1", replicasForscaled);
+                BinPackState.size = neededsized;
+                LastUpScaleDecision = Instant.now();
+                try (final KubernetesClient k8s = new KubernetesClientBuilder().build()) {
+                    k8s.apps().deployments().inNamespace("default").withName("latency").scale(neededsized);
+                    log.info("I have downscaled group {} you should have {}", "testgroup1", neededsized);
+                }
+                currentAssignment = assignment;
+            }
         }
         log.info("===================================");
     }
 
 
-    private static int binPackAndScale() {
+    private  static int binPackAndScale() {
         log.info(" shall we upscale group {}", "testgroup1");
         List<Consumer> consumers = new ArrayList<>();
         int consumerCount = 1;
@@ -70,18 +112,18 @@ public class BinPackState {
             if (partition.getLag() > mu*wsla*fup) {
                 log.info("Since partition {} has lag {} higher than consumer capacity times wsla {}" +
                         " we are truncating its lag", partition.getId(), partition.getLag(), mu*wsla*fup);
-                partition.setLag((long)(mu*wsla*fup/*dynamicAverageMaxConsumptionRate*wsla*/));
+                partition.setLag((long)(mu*wsla*fup));
             }
         }
         //if a certain partition has an arrival rate  higher than R  set its arrival rate  to R
         //that should not happen in a well partionned topic
         for (Partition partition : parts) {
-            if (partition.getArrivalRate() > mu*fup/*dynamicAverageMaxConsumptionRate*wsla*/) {
+            if (partition.getArrivalRate() > mu*fup) {
                 log.info("Since partition {} has arrival rate {} higher than consumer service rate {}" +
                                 " we are truncating its arrival rate", partition.getId(),
                         String.format("%.2f", partition.getArrivalRate()),
-                        String.format("%.2f", mu*fup /*dynamicAverageMaxConsumptionRate*wsla*/));
-                partition.setArrivalRate(mu*fup /*dynamicAverageMaxConsumptionRate*wsla*/);
+                        String.format("%.2f", mu*fup ));
+                partition.setArrivalRate(mu*fup );
             }
         }
         //start the bin pack FFD with sort
@@ -92,7 +134,7 @@ public class BinPackState {
             consumers.clear();
             for (int t = 0; t < consumerCount; t++) {
                 consumers.add(new Consumer((String.valueOf(t)),  (long)(mu*wsla*fup),
-                        mu*fup));
+                        mu*fup/*dynamicAverageMaxConsumptionRate*wsla*/));
             }
 
             for (j = 0; j < parts.size(); j++) {
@@ -113,18 +155,16 @@ public class BinPackState {
             if (j == parts.size())
                 break;
         }
-        assignment = consumers;
-        tempAssignment = consumers;
         log.info(" The BP up scaler recommended for group {} {}", "testgroup1", consumers.size());
         return consumers.size();
     }
 
-    static  int binPackAndScaled() {
+    private static  int binPackAndScaled() {
         log.info(" shall we down scale group {} ", "testgroup1");
         List<Consumer> consumers = new ArrayList<>();
         int consumerCount = 1;
-        List<Partition> parts = new ArrayList<>(ArrivalProducer.topicpartitions);
         float fdown = Float.valueOf(System.getenv("FDOWN"));
+        List<Partition> parts = new ArrayList<>(ArrivalProducer.topicpartitions);
         double fractiondynamicAverageMaxConsumptionRate = mu*fdown;
         for (Partition partition : parts) {
             if (partition.getLag() > fractiondynamicAverageMaxConsumptionRate*wsla) {
@@ -176,14 +216,22 @@ public class BinPackState {
             if (j == parts.size())
                 break;
         }
-
-        assignment = consumers;
         log.info(" The BP down scaler recommended  for group {} {}", "testgroup1", consumers.size());
         return consumers.size();
     }
 
-    private static boolean assignmentViolatesTheSLA2() {
-        float fup = Float.valueOf(System.getenv("FUP"));
+   /* private static  boolean assignmentViolatesTheSLA() {
+        for (Consumer cons : currentAssignment) {
+            if (cons.getRemainingLagCapacity() <  (long) (wsla*200*.9f)||
+                    cons.getRemainingArrivalCapacity() < 200f*0.9f){
+                return true;
+            }
+        }
+        return false;
+    }*/
+
+   /* private static boolean assignmentViolatesTheSLA2() {
+
         List<Partition> partsReset = new ArrayList<>(ArrivalProducer.topicpartitions);
         for (Consumer cons : currentAssignment) {
             double sumPartitionsArrival = 0;
@@ -193,12 +241,12 @@ public class BinPackState {
                 sumPartitionsLag += partsReset.get(p.getId()).getLag();
             }
 
-            if (sumPartitionsLag  > ( wsla*mu*fup)
-                    || sumPartitionsArrival > mu*fup) {
+            if (sumPartitionsLag  > ( wsla * 200  * .9f)
+                    || sumPartitionsArrival > 200* 0.9f) {
                 return true;
             }
         }
         return false;
-    }
+    }*/
 
 }
