+++ produceri3s/pom.xml
@@ -25,6 +25,12 @@
        <artifactId>kafka-clients</artifactId>
        <version>${kafka.version}</version>
    </dependency>
+       <dependency>
+           <groupId>org.example</groupId>
+           <artifactId>binscale-common</artifactId>
+           <version>1.0-SNAPSHOT</version>
+           <scope>compile</scope>
+       </dependency>
 
    </dependencies>
