--- produceri3s/src/main/java/fr/unice/scale/latencyaware/producer/CustomerSerializer.java
@@ -1,44 +1,0 @@
-package fr.unice.scale.latencyaware.producer;
-import org.apache.kafka.common.errors.SerializationException;
-import org.apache.kafka.common.serialization.Serializer;
-import fr.unice.scale.latencyaware.producer.Customer;
-
-import java.nio.ByteBuffer;
-import java.util.Map;
-
-public class CustomerSerializer implements Serializer<Customer> {
-    @Override
-    public void configure(Map configs, boolean isKey) {
-// nothing to configure
-    }
-    @Override
-    public byte[] serialize(String topic, Customer data) {
-        try {
-            byte[] serializedName;
-            int stringSize;
-            if (data == null)
-                return null;
-            else {
-                if (data.getName() != null) {
-                    serializedName = data.getName().getBytes("UTF-8");
-                    stringSize = serializedName.length;
-                } else {
-                    serializedName = new byte[0];
-                    stringSize = 0;
-                }
-            }
-            ByteBuffer buffer = ByteBuffer.allocate(4 + 4 + stringSize);
-            buffer.putInt(data.getID());
-            buffer.putInt(stringSize);
-            buffer.put(serializedName);
-            return buffer.array();
-        } catch (Exception e) {
-            throw new SerializationException(
-                    "Error when serializing Customer to byte[] " + e);
-        }
-    }
-    @Override
-    public void close() {
-// nothing to close
-    }
-}
