+++ produceri3s/src/main/java/fr/unice/scale/latencyaware/producer/workload/ConstantWorkload.java
@@ -1,6 +1,7 @@
-package fr.unice.scale.latencyaware.producer;
+package fr.unice.scale.latencyaware.producer.workload;
 
-import org.apache.kafka.clients.producer.KafkaProducer;
+import fr.unice.scale.latencyaware.common.entity.Customer;
+import fr.unice.scale.latencyaware.producer.KafkaProducerExample;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
