+++ produceri3s/src/main/java/fr/unice/scale/latencyaware/producer/workload/OldWorkload.java
@@ -1,9 +1,9 @@
-package fr.unice.scale.latencyaware.producer;
+package fr.unice.scale.latencyaware.producer.workload;
+import fr.unice.scale.latencyaware.common.entity.Customer;
+import fr.unice.scale.latencyaware.producer.KafkaProducerExample;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
-import fr.unice.scale.latencyaware.producer.Customer;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
