+++ produceri3s/src/main/java/fr/unice/scale/latencyaware/producer/workload/OldWorkloadSkewed.java
@@ -1,4 +1,6 @@
-package fr.unice.scale.latencyaware.producer;
+package fr.unice.scale.latencyaware.producer.workload;
+import fr.unice.scale.latencyaware.common.entity.Customer;
+import fr.unice.scale.latencyaware.producer.KafkaProducerExample;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -11,7 +13,7 @@ import java.util.UUID;
 
 public class OldWorkloadSkewed {
     static float ArrivalRate;
-    public static void  startWorkload() throws IOException, URISyntaxException, InterruptedException {
+    public static void startWorkload() throws IOException, URISyntaxException, InterruptedException {
 
         final Logger log = LogManager.getLogger(OldWorkloadSkewed.class);
         Workload wrld = new Workload();
