+++ produceri3s/src/main/java/fr/unice/scale/latencyaware/producer/workload/Workload.java
@@ -1,4 +1,4 @@
-package fr.unice.scale.latencyaware.producer;
+package fr.unice.scale.latencyaware.producer.workload;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
