+++ variableconsumer/src/main/java/fr/unice/scale/latencyaware/consumer/assignor/AbstractAssignor.java
@@ -1,4 +1,4 @@
-package fr.unice.scale.latencyaware.consumer;
+package fr.unice.scale.latencyaware.consumer.assignor;
 import org.apache.kafka.clients.consumer.ConsumerPartitionAssignor;
 import org.apache.kafka.common.Cluster;
 import org.apache.kafka.common.TopicPartition;
