+++ variableconsumer/src/main/java/fr/unice/scale/latencyaware/consumer/metrics/TimeMeasure.java
@@ -1,4 +1,4 @@
-package fr.unice.scale.latencyaware.consumer;
+package fr.unice.scale.latencyaware.consumer.metrics;
 
 public class TimeMeasure {
     double duration;
